backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.7
nltk==3.8.1
docker==7.0.0
pypiwin32==223
pandas==2.0.0
requests==2.31.0
selenium==4.18.1
langchain-openai==0.0.5
playwright==1.41.2
transformers==4.37.2
torch==2.2.0
Pillow==10.2.0



